Rename LocationProvider to LocationResolver
